deploy: draft catalyst contract deploy setup
