+++ packages/deploy/deploy/300_catalyst/300_deploy_operator_registrant.ts
@@ -1,0 +1,19 @@
+import { DeployFunction } from "hardhat-deploy/types";
+import { HardhatRuntimeEnvironment } from "hardhat/types";
+
+const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
+  const { deployments, getNamedAccounts } = hre;
+  const { deploy } = deployments;
+
+  const { deployer } = await getNamedAccounts();
+
+  // Operator filter subscription
+  await deploy("OperatorFilterRegistrant", {
+    from: deployer,
+    contract: "@sandbox-smart-contracts/asset/contracts/OperatorFilter/OperatorFilterRegistrant.sol:OperatorFilterRegistrant",
+    log: true,
+    skipIfAlreadyDeployed: true,
+  });
+};
+export default func;
+func.tags = ["OperatorFilterRegistrant", 'L2'];
