+++ packages/deploy/hardhat.config.ts
@@ -8,7 +8,8 @@ import './tasks/importedPackages';
 
 // Package name : solidity source code path
 const importedPackages = {
-  '@sandbox-smart-contracts/giveaway': 'contracts/SignedMultiGiveaway.sol',
+  '@sandbox-smart-contracts/asset': 'contracts/',
+  '@sandbox-smart-contracts/giveaway': 'contracts/SignedMultiGiveaway.sol'
 };
 
 const namedAccounts = {
